+++ utils/airsim_manager.py
@@ -136,36 +136,6 @@ class AirsimManager:
         }
         return car_position_and_speed
 
-    # def get_car1_state(self,):
-    #     car1_position_and_speed = self.get_car_position_and_speed(self.experiment.CAR1_NAME)
-    #     car2_position_and_speed = self.get_car_position_and_speed(self.experiment.CAR2_NAME)
-    #     car1_state = np.array([
-    #         car1_position_and_speed["x"],
-    #         car1_position_and_speed["y"],
-    #         abs(car1_position_and_speed["Vx"]),
-    #         abs(car1_position_and_speed["Vy"]),
-    #     ])
-    #     #
-    #     # if logger is not None and self.experiment.LOG_CAR_STATES:
-    #     #     logger.log_state(car1_state, self.experiment.CAR1_NAME)
-    #     return car1_state
-    #
-    # def get_car2_state(self):
-    #     car2_position_and_speed = self.get_car_position_and_speed(self.experiment.CAR2_NAME)
-    #     car1_position_and_speed = self.get_car_position_and_speed(self.experiment.CAR1_NAME)
-    #     car2_state = np.array([
-    #         car2_position_and_speed["x"],
-    #         car2_position_and_speed["y"],
-    #         abs(car2_position_and_speed["Vx"]),
-    #         abs(car2_position_and_speed["Vy"]),
-    #     ])
-    #     #print(car2_position_and_speed["x"],car2_position_and_speed["y"])
-    #
-    #     # if logger is not None and self.experiment.LOG_CAR_STATES:
-    #     #     logger.log_state(car2_state, self.experiment.CAR2_NAME)
-    #
-    #     return car2_state
-
     def get_car1_state(self):
         pose = self.airsim_client.simGetObjectPose(self.experiment.CAR1_NAME)
         position = pose.position
