+++ utils/plotting_utils.py
@@ -40,3 +40,36 @@ class PlottingUtils:
         plt.ylabel('Action')
         plt.legend()
 
+def plot_metrics(metrics_path):
+    metrics = pd.read_csv(metrics_path)
+    plt.figure(figsize=(10, 6))
+    plt.plot(metrics["loss"], label="Loss")
+    plt.title("Loss over Episodes")
+    plt.xlabel("Episode")
+    plt.ylabel("Loss")
+    plt.legend()
+    plt.grid()
+    plt.show()
+    plt.figure(figsize=(10, 6))
+    plt.plot(metrics["entropy"], label="Entropy", color="orange")
+    plt.title("Entropy over Episodes")
+    plt.xlabel("Episode")
+    plt.ylabel("Entropy")
+    plt.legend()
+    plt.grid()
+    plt.show()
+    plt.figure(figsize=(10, 6))
+    plt.plot(metrics["return"], label="Return", color="green")
+    plt.title("Return over Episodes")
+    plt.xlabel("Episode")
+    plt.ylabel("Return")
+    plt.legend()
+    plt.grid()
+    plt.show()
+
+
+def plot_results(path, all_rewards, all_actions):
+    PlottingUtils.plot_losses(path)
+    PlottingUtils.plot_rewards(all_rewards)
+    PlottingUtils.show_plots()
+    PlottingUtils.plot_actions(all_actions)
