+++ utils/training_loop.py
@@ -9,7 +9,7 @@ import os
 from utils.model.model_handler import Model
 from utils.agent_handler import Agent
 from utils.airsim_manager import AirsimManager
-from utils.plotting_utils import PlottingUtils
+from utils.plotting_utils import *
 
 def inference_loop(p_agent_loss, p_master_loss, p_episode_counter, experiment, env, agent_model, master_model, agent):
     """
@@ -23,20 +23,19 @@ def inference_loop(p_agent_loss, p_master_loss, p_episode_counter, experiment, e
     """
     collision_counter, episode_counter, total_steps = 0, 0, 0
     all_rewards, all_actions = [], []
-    if experiment.ONLY_INFERENCE: # inference mode logic only
-        for episode in range(experiment.EPISODES_PER_CYCLE*experiment.CYCLES):
-            episode_counter += 1
-            p_episode_counter.append(episode_counter)
-            print(f"  @ Episode {episode_counter} INFERENCE @")
-            # Run an inference episode that uses separate agent models for each vehicle.
-            episode_rewards, episode_actions, steps, p_agent_loss, p_master_loss, episode_states = run_episode_inference(
-                env, agent_model['agent1'], agent_model['agent2'], master_model, experiment
-            )
-            total_steps += steps
-            all_rewards.append(episode_rewards)
-            all_actions.append(episode_actions)
-        print("Inference completed.")
-        return p_episode_counter, p_agent_loss, p_master_loss, agent_model, collision_counter, all_rewards, all_actions
+    for episode in range(experiment.EPISODES_PER_CYCLE*experiment.CYCLES):
+        episode_counter += 1
+        p_episode_counter.append(episode_counter)
+        print(f"  @ Episode {episode_counter} INFERENCE @")
+        # Run an inference episode that uses separate agent models for each vehicle.
+        episode_rewards, episode_actions, steps, p_agent_loss, p_master_loss, episode_states = run_episode_inference(
+            env, agent_model['agent1'], agent_model['agent2'], master_model, experiment
+        )
+        total_steps += steps
+        all_rewards.append(episode_rewards)
+        all_actions.append(episode_actions)
+    print("Inference completed.")
+    return p_episode_counter, p_agent_loss, p_master_loss, agent_model, collision_counter, all_rewards, all_actions
 
 def run_episode_inference(env, agent1_model, agent2_model, master_model, experiment):
     """
@@ -56,14 +55,9 @@ def run_episode_inference(env, agent1_model, agent2_model, master_model, experim
       p_master_loss: (Empty list in inference mode)
       all_states: List of master inputs from each step.
     """
-    episode_sum = 0
-    steps_counter = 0
-    actions_list = []
-    all_states = []
-    p_agent_loss = []
-    p_master_loss = []
+    episode_sum, steps_counter = 0, 0
+    actions_list, all_states, p_agent_loss, p_master_loss = [], [], [], []
     resume_experiment_simulation(env)
-
     while True:
         steps_counter += 1
         # Retrieve local states for Car1 and Car2 (each 4-dim)
@@ -101,53 +95,52 @@ def run_episode_inference(env, agent1_model, agent2_model, master_model, experim
 # Main Experiment Runner
 # -------------------------------
 def run_experiment(experiment_config):
-    if experiment_config.ONLY_INFERENCE:
-        p_agent_loss = []
-        p_master_loss = []
-        p_episode_counter = []
-        # Initialize AirSim Manager
-        airsim_manager = AirsimManager(experiment_config)
-        # Initialize Master Model and load pre-trained weights
-        master_model = MasterModel(embedding_size=4, airsim_manager=airsim_manager, experiment=experiment_config)
-        master_model.load("EXP5_Inference_Models/trained_model_master.zip")
-        # Create two separate Agent instances (one for each car) with the same master model
-        agent1 = Agent(experiment_config, airsim_manager, master_model)
-        agent2 = Agent(experiment_config, airsim_manager, master_model)
-        # Create the shared environment
-        env = DummyVecEnv([lambda: Agent(experiment_config, airsim_manager, master_model)])
-        # Load separate pre-trained agent models for each vehicle
-        agent1_model = Model.load("EXP5_Inference_Models/agent_trained_model.zip", env,
-                                  experiment_config).model
-        agent2_model = Model.load("EXP5_Inference_Models/agent_trained_model.zip", env,
-                                  experiment_config).model
-        # Configure loggers
-        base_path = experiment_config.EXPERIMENT_PATH
-        agent_logger_path = os.path.join(base_path, "agent_logs")
-        agent_logger = configure(agent_logger_path, ["stdout", "csv", "tensorboard"])
-        master_logger_path = os.path.join(base_path, "master_logs")
-        master_model_logger = configure(master_logger_path, ["stdout", "csv", "tensorboard"])
-        agent1_model.set_logger(agent_logger)
-        master_model.set_logger(master_model_logger)
-        # Create a dictionary for the agent models
-        agent_models = {'agent1': agent1_model, 'agent2': agent2_model}
-        # Run the inference loop; note that the inference branch uses the two-agent-model version.
-        p_episode_counter, p_agent_loss, p_master_loss, _, collision_counter, all_rewards, all_actions = inference_loop(
-            p_agent_loss, p_master_loss, p_episode_counter,
-            experiment=experiment_config,
-            env=env,
-            agent_model=agent_models,  # dictionary with keys 'agent1' and 'agent2'
-            master_model=master_model,
-            agent=agent1  # Pass one of the agent instances; the env is shared.
-        )
-        print(p_master_loss)
-        # Optionally re-save the loaded models
-        agent1_model.save(experiment_config.SAVE_MODEL_DIRECTORY + "_agent1_inference")
-        agent2_model.save(experiment_config.SAVE_MODEL_DIRECTORY + "_agent2_inference")
-        master_model.save(experiment_config.SAVE_MODEL_DIRECTORY + "_master_inference")
-        agent_logger.close()
-        master_model_logger.close()
-        print("Models loaded and inference completed")
-        print("Total collisions:", collision_counter)
+    p_agent_loss = []
+    p_master_loss = []
+    p_episode_counter = []
+    # Initialize AirSim Manager
+    airsim_manager = AirsimManager(experiment_config)
+    # Initialize Master Model and load pre-trained weights
+    master_model = MasterModel(embedding_size=experiment_config.EMBEDDING_SIZE, airsim_manager=airsim_manager, experiment=experiment_config)
+    master_model.load(experiment_config.MASTER_TRAINED_MODEL)
+    # Create two separate Agent instances (one for each car) with the same master model
+    agent1 = Agent(experiment_config, airsim_manager, master_model)
+    # Create the shared environment
+    env = DummyVecEnv([lambda: Agent(experiment_config, airsim_manager, master_model)])
+    # Load separate pre-trained agent models for each vehicle
+    agent1_model = Model.load(experiment_config.AGENT_TRAINED_MODEL, env,
+                              experiment_config).model
+    agent2_model = Model.load(experiment_config.AGENT_TRAINED_MODEL, env,
+                              experiment_config).model
+    # Configure loggers
+    base_path = experiment_config.EXPERIMENT_PATH
+    agent_logger_path = os.path.join(base_path, "agent_logs")
+    agent_logger = configure(agent_logger_path, ["stdout", "csv", "tensorboard"])
+    master_logger_path = os.path.join(base_path, "master_logs")
+    master_model_logger = configure(master_logger_path, ["stdout", "csv", "tensorboard"])
+    agent1_model.set_logger(agent_logger)
+    master_model.set_logger(master_model_logger)
+    # Create a dictionary for the agent models
+    agent_models = {'agent1': agent1_model, 'agent2': agent2_model}
+    # Run the inference loop; note that the inference branch uses the two-agent-model version.
+    #all arrays are passing from and to functions, marked with p (ie p_episode_counter)
+    p_episode_counter, p_agent_loss, p_master_loss, _, collision_counter, all_rewards, all_actions = inference_loop(
+        p_agent_loss, p_master_loss, p_episode_counter,
+        experiment=experiment_config,
+        env=env,
+        agent_model=agent_models,  # dictionary with keys 'agent1' and 'agent2'
+        master_model=master_model,
+        agent=agent1  # Pass one of the agent instances; the env is shared.
+    )
+    print(p_master_loss)
+    # Optionally re-save the loaded models
+    agent1_model.save(experiment_config.SAVE_MODEL_DIRECTORY + "_agent1_inference")
+    agent2_model.save(experiment_config.SAVE_MODEL_DIRECTORY + "_agent2_inference")
+    master_model.save(experiment_config.SAVE_MODEL_DIRECTORY + "_master_inference")
+    agent_logger.close()
+    master_model_logger.close()
+    print("Models loaded and inference completed")
+    print("Total collisions:", collision_counter)
 # -------------------------------
 # Override GYM Environment Functions
 # -------------------------------
@@ -157,45 +150,3 @@ def resume_experiment_simulation(env):
 
 def pause_experiment_simulation(env):
     env.envs[0].airsim_manager.pause_simulation()
-
-
-
-def plot_results(path, all_rewards, all_actions):
-    PlottingUtils.plot_losses(path)
-    PlottingUtils.plot_rewards(all_rewards)
-    PlottingUtils.show_plots()
-    PlottingUtils.plot_actions(all_actions)
-
-def resume_experiment_simulation(env):
-    env.envs[0].airsim_manager.resume_simulation()
-
-
-def pause_experiment_simulation(env):
-    env.envs[0].airsim_manager.pause_simulation()
-
-def plot_metrics(metrics_path):
-    metrics = pd.read_csv(metrics_path)
-    plt.figure(figsize=(10, 6))
-    plt.plot(metrics["loss"], label="Loss")
-    plt.title("Loss over Episodes")
-    plt.xlabel("Episode")
-    plt.ylabel("Loss")
-    plt.legend()
-    plt.grid()
-    plt.show()
-    plt.figure(figsize=(10, 6))
-    plt.plot(metrics["entropy"], label="Entropy", color="orange")
-    plt.title("Entropy over Episodes")
-    plt.xlabel("Episode")
-    plt.ylabel("Entropy")
-    plt.legend()
-    plt.grid()
-    plt.show()
-    plt.figure(figsize=(10, 6))
-    plt.plot(metrics["return"], label="Return", color="green")
-    plt.title("Return over Episodes")
-    plt.xlabel("Episode")
-    plt.ylabel("Return")
-    plt.legend()
-    plt.grid()
-    plt.show()
